+++ src/index.ts
@@ -1,4 +1,4 @@
 import { Auditor } from "./core";
-import { TAuditOptions, TEvent, TFileConfig } from "./types"
+import { TAuditOptions, TEvent, TFileConfig } from "./types";
 
 export { Auditor, TAuditOptions, TEvent, TFileConfig };